Added deleting images
