+++ src/EducationService.Data/Interfaces/IUserEducationRepository.cs
@@ -1,6 +1,7 @@
 ﻿using LT.DigitalOffice.EducationService.Models.Db;
 using LT.DigitalOffice.EducationService.Models.Dto.Requests.User;
 using LT.DigitalOffice.Kernel.Attributes;
+using Microsoft.AspNetCore.JsonPatch;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
@@ -10,9 +11,15 @@ namespace LT.DigitalOffice.EducationService.Data.Interfaces
   [AutoInject]
   public interface IUserEducationRepository
   {
-    Task<List<DbUserEducation>> GetAsync(Guid userId);
+    Task<Guid?> CreateAsync(DbUserEducation dbEducation);
 
-    Task DisactivateEducationsAsync(Guid userId, Guid modifiedBy);
+    Task<DbUserEducation> GetAsync(Guid educationId);
+
+    Task<bool> EditAsync(DbUserEducation educationId, JsonPatchDocument<DbUserEducation> request);
+
+    Task<bool> RemoveAsync(DbUserEducation dbEducation);
+
+    Task<List<Guid>> DisactivateEducationsAsync(Guid userId, Guid modifiedBy);
 
     Task<List<DbUserEducation>> FindAsync(FindUsersFilter filter);
   }
