--- src/EducationService.Mappers/Models/Interfaces/IUserEducationDataMapper.cs
@@ -1,12 +1,0 @@
-﻿using LT.DigitalOffice.EducationService.Models.Db;
-using LT.DigitalOffice.Kernel.Attributes;
-using LT.DigitalOffice.Models.Broker.Models.Education;
-
-namespace LT.DigitalOffice.EducationService.Mappers.Models.Interfaces
-{
-  [AutoInject]
-  public interface IUserEducationDataMapper
-  {
-    EducationData Map(DbUserEducation dbUserEducation);
-  }
-}
