+++ src/EducationService/appsettings.json
@@ -31,8 +31,8 @@
     //Common
     "DisactivateUserEndpoint": "Common_DisactivateUser",
 
-    //Education
-    "DisactivateUserEducationsEndpoint": "EducationService_DisactivateUser",
+    //Subscriptions
+    "DisactivateUserEducationsEndpoint": "Common_DisactivateUserEducations",
 
     // CRService
     "CheckUserRightsEndpoint": "CRService_CheckUserRights",
